+++ src/renderer/src/pages/override.tsx
@@ -164,7 +164,7 @@ const Override: React.FC = () => {
             variant="light"
             size="sm"
             onPress={() => {
-              open('https://github.com/mihomo-party-org/override-hub')
+              open('https://github.com/markd3ng/override-hub')
             }}
           >
             <RiArchiveLine className="text-lg" />